Refactor into separate modules and update docs
